aws: do not timeout if there is a violation, instead just return an error
This is a false positive and causes us too timeout when we are calling
destroy on a tf file, which is used by instances outside the terraform
state file.  In that case, because the instances can't be created
anymore, Terraform doesn't return an error in the case of a
DependencyViolation, instead it's tries to delete it over and over
again, until it times out.

We wan't eventually that it returns immediately, so we know what the
underlying problem is. Right now it's being hided.