vendor: remove hashicorp/hil
This is temporary workaround for koding/koding@8375 - due to lack of
vendor tree flattening stackplan/template uses different hil package
than the terraform does.

This should be fixed by flattening the vendor tree, however it is not
that trivial, as there is a number of conflicting versions
(non-backward-compatible), which should be resolved first before
flattening.

Removing this vendored copy works like manual flattening of the hil
package.